custom-exporter/requirements.txt
pymongo
flask
mysql-connector-python